REFACTOR: loadUser 상속 메서드
