+++ backend/src/main/java/org/ftclub/cabinet/security/handler/CustomOAuth2UserService.java
@@ -3,7 +3,7 @@ package org.ftclub.cabinet.security.handler;
 import java.util.Map;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.ftclub.cabinet.auth.domain.CustomOauth2User;
+import org.ftclub.cabinet.oauth.domain.CustomOauth2User;
 import org.springframework.security.oauth2.client.userinfo.DefaultOAuth2UserService;
 import org.springframework.security.oauth2.client.userinfo.OAuth2UserRequest;
 import org.springframework.security.oauth2.core.OAuth2AuthenticationException;
